+++ tech1-framework-domain/src/main/java/io/tech1/framework/domain/geo/GeoCountryFlag.java
@@ -1,0 +1,12 @@
+package io.tech1.framework.domain.geo;
+
+import lombok.Data;
+
+// Lombok
+@Data
+public class GeoCountryFlag {
+    private final String name;
+    private final String code;
+    private final String emoji;
+    private final String unicode;
+}