+++ tech1-framework-utilities/src/main/java/io/tech1/framework/utilities/geo/facades/GeoCountryFlagUtility.java
@@ -1,0 +1,6 @@
+package io.tech1.framework.utilities.geo.facades;
+
+public interface GeoCountryFlagUtility {
+    String getEmojiByCountry(String country);
+    String getEmojiByCountryCode(String countryCode);
+}